Fix two bugs caused by integer truncation by more rigorous casting.  One meant 64-bit LE values were de-serialized incorrectly, the other that large partitions were effectively truncated when accessed as a stream.

+++ src/Partitions/BiosPartitionTable.cs
@@ -413,7 +413,7 @@ namespace DiscUtils.Partitions
 
         internal Stream Open(BiosPartitionRecord record)
         {
-            return new SubStream(_diskData, false, record.LBAStartAbsolute * Utilities.SectorSize, record.LBALength * Utilities.SectorSize);
+            return new SubStream(_diskData, false, ((long)record.LBAStartAbsolute) * Utilities.SectorSize, ((long)record.LBALength) * Utilities.SectorSize);
         }
     }
 }
